+++ functions/Get-PSReminderTag.ps1
@@ -3,41 +3,39 @@
 Function Get-PSReminderTag {
     [cmdletbinding()]
     [OutputType('PSReminderTag')]
+    [alias('gprt')]
     Param()
 
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Starting $($MyInvocation.MyCommand)"
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running under PowerShell version $($PSVersionTable.PSVersion)"
+        $PSDefaultParameterValues['_verbose:Command'] = $MyInvocation.MyCommand
+        $PSDefaultParameterValues['_verbose:block'] = 'Begin'
+        _verbose $($strings.Starting -f $($MyInvocation.MyCommand))
+        _verbose $($strings.PSVersion -f $($PSVersionTable.PSVersion))
     } #begin
 
     Process {
-        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Getting PSReminder Tags from ``$PSReminderTag"
-        #TODO get tags from the database not defined here
-        $PSReminderTag.GetEnumerator() | ForEach-Object {
-            $stringValue = $_.Value.Replace("$([char]27)", "``e")
-            [PSCustomObject]@{
-                PSTypeName = 'PSReminderTag'
-                Tag        = $_.Key
-                Style      = '{0}{1}{2}' -f $($_.Value), $stringValue, $("`e[0m")
-            }
-        }
+        $PSDefaultParameterValues['_verbose:block'] = 'Process'
+        _verbose $($strings.GetTags)
+        $pref = Get-PSReminderPreference
+        $pref.ShowTags()
 
-        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Getting defined tags from the database not in `$PSReminderTag"
-        (Invoke-MySQLiteQuery -query "select tags from $PSReminderTable where tags like '%'" -Path $PSReminderDB).foreach({$_.tags.split(",")}) |
-        Select-Object -unique |
-        Where-Object {$PSReminderTag.keys -notContains $_} |
-        Foreach-Object {
+        _verbose $($strings.GetUndefinedTags)
+        (Invoke-MySQLiteQuery -Query "select tags from $PSReminderTable where tags like '%'" -Path $PSReminderDB).foreach({ $_.tags.split(',') }) |
+        Select-Object -Unique |
+        Where-Object { $PSReminderTag.keys -notContains $_ } |
+        ForEach-Object {
             [PSCustomObject]@{
                 PSTypeName = 'PSReminderTag'
                 Tag        = $_
                 Style      = $Null
             }
         }
-
     } #process
 
     End {
-        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Ending $($MyInvocation.MyCommand)"
+        $PSDefaultParameterValues['_verbose:Command'] = $MyInvocation.MyCommand
+        $PSDefaultParameterValues['_verbose:block'] = 'End'
+        _verbose $($strings.Ending -f $($MyInvocation.MyCommand))
     } #end
 
 } #close Get-PSReminderTag