+++ functions/Import-FromTickleDatabase.ps1
@@ -2,7 +2,7 @@
 
 Function Import-FromTickleDatabase {
     [cmdletbinding(SupportsShouldProcess)]
-    [OutputType('Nsone')]
+    [OutputType('None')]
     Param(
         [Parameter(
             Position = 0,
@@ -23,22 +23,25 @@ Function Import-FromTickleDatabase {
     )
 
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Starting $($MyInvocation.MyCommand)"
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] Running under PowerShell version $($PSVersionTable.PSVersion)"
+        $PSDefaultParameterValues['_verbose:Command'] = $MyInvocation.MyCommand
+        $PSDefaultParameterValues['_verbose:block'] = 'Begin'
+        _verbose $($strings.Starting -f $($MyInvocation.MyCommand))
+        _verbose $($strings.PSVersion -f $($PSVersionTable.PSVersion))
         $InvokeParams = @{
             Query       = $null
             KeepAlive   = $True
             ErrorAction = 'Stop'
         }
         $ProgParams = @{
-            Activity = "Importing TickleData"
-            Status   = ""
-            PercentComplete =0
+            Activity        = $strings.ImportTD
+            Status          = ''
+            PercentComplete = 0
         }
     } #begin
 
     Process {
-        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Importing Tickle data from $Path"
+        $PSDefaultParameterValues['_verbose:block'] = 'Process'
+        _verbose $($strings.ImportTDPath -f $Path)
         Try {
             $in = Import-Clixml -Path $Path -ErrorAction Stop
         }
@@ -48,7 +51,7 @@ Function Import-FromTickleDatabase {
         If ($In.count -gt 0) {
             #Initialize a new database
             If (Test-Path -Path $DatabasePath) {
-                Write-Warning "The database file $DatabasePath already exists. Please remove it or specify a different path."
+                Write-Warning $($strings.DBExists -f $DatabasePath)
                 return
             }
             else {
@@ -56,11 +59,10 @@ Function Import-FromTickleDatabase {
 
                 #open connection
                 $conn = Open-MySQLiteDB -Path $DatabasePath
-                $InvokeParams.Add("Connection", $conn)
-                $global:ip = $InvokeParams
+                $InvokeParams.Add('Connection', $conn)
             }
 
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Importing $($In.Count) events"
+            _verbose $($strings.ImportCount -f $($In.Count))
             #initialize a counter
             $i = 0
             foreach ($item in $in) {
@@ -70,12 +72,12 @@ Function Import-FromTickleDatabase {
                 $evt = $item.EventName.Replace("'", '')
 
                 $i++
-                $ProgParams.PercentComplete = $i/$in.count*100
-                $ProgParams.Status = "Importing [$($item.EventID)] $($item.EventName)"
+                $ProgParams.PercentComplete = $i / $in.count * 100
+                $ProgParams.Status = $strings.ImportTickle -f $($item.EventID). $($item.EventName)
                 Write-Progress @ProgParams
                 If ($Item.Archived) {
                     #Import events where archived into the ArchiveData table
-                    Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Archiving event [$($item.EventID)] $($Item.EventName)"
+                    _verbose $($strings.ArchiveTickle -f $($Item.EventID), $($Item.EventName))
                     If ($PSCmdlet.ShouldProcess("[$($item.EventID)] $($Item.EventName)", "`e[38;5;222mArchive event`e[0m")) {
                         #Archive the event
                         $InvokeParams.Query = "Insert into $PSReminderArchiveTable (EventID,EventName,EventDate,EventComment,Tags,ArchivedDate) values ('$($item.EventID)','$($evt)','$($evtDate)','$($item.EventComment)','','$((Get-Date).ToString('yyyy-MM-dd HH:mm:ss'))')"
@@ -85,7 +87,8 @@ Function Import-FromTickleDatabase {
                 } #if archived
                 else {
                     #Import events where not archived into the EventData table
-                    Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] Importing event [$($item.EventID)] $($item.Name)"
+                    _verbose $($strings.ImportTickle -f $($Item.EventID), $($Item.EventName))
+
                     If ($PSCmdlet.ShouldProcess("[$($item.EventID)] $($Item.EventName)", 'Import event')) {
                         #Add the event
                         $InvokeParams.Query = "Insert into $PSReminderTable (EventID,EventName,EventDate,EventComment,Tags) values ('$($item.EventID)','$($evt)','$($evtDate)','$($item.EventComment)','')"
@@ -99,11 +102,13 @@ Function Import-FromTickleDatabase {
     } #process
 
     End {
+        $PSDefaultParameterValues['_verbose:Command'] = $MyInvocation.MyCommand
+        $PSDefaultParameterValues['_verbose:block'] = 'End'
         if ($conn.State -eq 'Open') {
-            Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Closing the database connection"
+            _verbose $($strings.CloseDB)
             $conn.Close()
         }
-        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] Ending $($MyInvocation.MyCommand)"
+        _verbose $($strings.Ending -f $($MyInvocation.MyCommand))
     } #end
 
 } #close Import-FromTickleDatabase