+++ functions/private.ps1
@@ -18,11 +18,14 @@ function _NewPSReminder {
         [Alias('Comment')]
         [String]$EventComment,
         [Parameter(ValueFromPipelineByPropertyName)]
-        [String]$Tags
+        [String]$Tags,
+        [string]$Source
     )
     Process {
         [string[]]$TagArray = $Tags -split ',' | Foreach-Object {$_.Trim()}
-        New-Object -TypeName PSReminder -ArgumentList @($eventID, $EventName, $EventDate, $EventComment,$TagArray)
+        $obj = New-Object -TypeName PSReminder -ArgumentList @($eventID, $EventName, $EventDate, $EventComment,$TagArray)
+        $obj.Source = $Source
+        $obj
     }
 } #close _NewPSReminder
 
@@ -46,10 +49,37 @@ function _NewArchivePSReminder {
         [Parameter(ValueFromPipelineByPropertyName)]
         [String]$Tags,
         [Parameter(ValueFromPipelineByPropertyName)]
-        [DateTime]$ArchivedDate
+        [DateTime]$ArchivedDate,
+        [string]$Source
     )
     Process {
         [string[]]$TagArray = $Tags -split ',' | Foreach-Object {$_.Trim()}
-        New-Object -TypeName ArchivePSReminder -ArgumentList @($eventID, $EventName, $EventDate, $EventComment,$TagArray,$ArchivedDate)
+        $obj = New-Object -TypeName ArchivePSReminder -ArgumentList @($eventID, $EventName, $EventDate, $EventComment,$TagArray,$ArchivedDate)
+        $obj.Source = $Source
+        $obj
     }
 } #close _NewPSReminder
+
+#my custom verbose function
+function _verbose {
+    [CmdletBinding()]
+    param(
+        [Parameter(Position = 0)]
+        [string]$Message,
+        [string]$Block = 'PROCESS',
+        [string]$Command
+    )
+
+    [string]$ANSI =  "`e[48;5;226;38;5;232m"
+
+    $BlockString = $Block.ToUpper().PadRight(7, ' ')
+    $Reset = "$([char]27)[0m"
+    $ToD = (Get-Date).TimeOfDay
+    $AnsiCommand = "$([char]27)$Ansi$($command)"
+    $Italic = "$([char]27)[3m"
+    #Write-Verbose "[$((Get-Date).TimeOfDay) $BlockString] $([char]27)$Ansi$($command)$([char]27)[0m: $([char]27)[3m$message$([char]27)[0m"
+    $msg = '[{0} {1}] {2}{3}-> {4} {5}{3}' -f $Tod, $BlockString, $AnsiCommand, $Reset, $Italic, $Message
+
+    Write-Verbose -Message $msg
+
+}